newfan_analysis/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.2"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
sqlalchemy = "^2.0.33"
aiomysql = "^0.2.0"
langchain = "^0.2.16"
boto3 = "^1.35.11"
pysqlite3 = "^0.5.3"
pandas = "^2.2.2"
swifter = "^1.4.0"
langchain-experimental = "^0.0.65"
langchain-community = "^0.2.16"
mysql-connector-python = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
